Move AlertDialog and BalanceSortDialog to ui/dialogs package
